refactor: move client integration tests to loon-testkit
Move all 28 integration test files from loon-client/tests/ to
loon-testkit/tests/ so that loon-client has no dependency on
loon-server at all (not even dev). This fixes CI OOM failures
caused by each test binary linking the full server dependency tree.

Also moves the file.rs #[cfg(test)] block to loon-testkit/tests/ops_file.rs
and reverts the CARGO_BUILD_JOBS=1 CI workaround since it's no longer needed.
