+++ .github/workflows/ci.yml
@@ -37,8 +37,6 @@ jobs:
   test:
     name: test
     runs-on: ubuntu-latest
-    env:
-      CARGO_BUILD_JOBS: 1
     steps:
       - uses: actions/checkout@v4
       - uses: actions/cache@v4
@@ -49,6 +47,4 @@ jobs:
             target
           key: ${{ runner.os }}-cargo-test-${{ hashFiles('**/Cargo.lock') }}
           restore-keys: ${{ runner.os }}-cargo-test-
-      - name: Free disk space
-        run: df -h / && rm -rf target/debug/incremental && echo "--- after cleanup ---" && df -h /
       - run: cargo test --all
