+++ crates/loon-testkit/Cargo.toml
@@ -11,9 +11,9 @@ anyhow.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_yaml.workspace = true
+thiserror.workspace = true
+toml.workspace = true
+loon-client = { path = "../loon-client" }
 loon-server = { path = "../loon-server" }
 loon-sim = { path = "../loon-sim" }
 loon-types = { path = "../loon-types" }
-
-[dev-dependencies]
-loon-client = { path = "../loon-client" }
