+++ crates/loon-testkit/tests/client_observe_delete_move.rs
