+++ crates/loon-testkit/tests/client_remote_only_discovery.rs
