+++ crates/loon-testkit/tests/local_only_binding.rs
